UI: Move methods into MapNode class
Moved the following methods:
- calculateNodeGroup
- getGroupOperatingSystem
- getLabel
